pyproject.toml
[tool.poetry]
name = "projetodjango"
version = "0.1.0"
description = ""
authors = ["Pedro Matumoto <pedro.matumoto@scicrop.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
